+++ packages/tanstack_query/test/refetch_on_events_test.dart
@@ -8,8 +8,9 @@ void main() {
 
   setUp(() {
     queryClient = QueryClient(
-        defaultOptions:
-            const DefaultOptions(queries: QueryDefaultOptions(gcTime: 0)));
+        defaultOptions: const DefaultOptions(
+            queries: QueryDefaultOptions(gcTime: -1),
+            mutations: MutationDefaultOptions(gcTime: -1)));
     queryClient.queryCache.clear();
 
     // ensure managers are in a default state